Rename 'components/DataViewBasicInfoPanel.vue' to 'components/index/_shared/DataView/BasicInfoPanel.vue'
