+++ components/index/_shared/DataView/BasicInfoPanel.vue
